+++ src/core/CommandLoader.ts
@@ -1,7 +1,6 @@
-import TEV from "../Main";
+import TEV from "../main";
 import fs from "node:fs";
 import path from "node:path";
-import { red } from "../modules/Colors";
 import CommandExecutor from "./CommandExecutor";
 
 export default class CommandLoader {
@@ -10,15 +9,12 @@ export default class CommandLoader {
     public async loadMain(main: TEV): Promise<void> {
         this.tev = main;
 
-        // ? Checks if there's a main file.
         if (!("main" in this.tev.config.content)) {
-            throw new Error(`${red(main.text)} - Main is a required argument.`);
-            process.exit(1);
+            throw new Error("[TEV - Error] - Main is a required argument.");
         }
 
-        // ? commands is the name given to the file content including a code made from "tev" functions.
-        const commands = fs.readFileSync(path.resolve(process.cwd(), this.tev.config.content.main), "utf-8");
-        const executor = new CommandExecutor()
-        await executor.execute(commands.toString().trim().split("\n"));
+        const commands = fs.readFileSync(path.resolve(process.cwd(), this.tev.config.content.main), "utf-8").trim();
+        const executor = new CommandExecutor(this.tev);
+        await executor.execute(commands.split("\n"));
     }
 }