+++ src/core/utils/comments.ts
@@ -1,17 +1,13 @@
 import { ArgTypes, Types } from "../../types/enums";
 
 export const $stripeComment = {
-    type: Types.METHOD,
+    name: "$stripeComment",
     lib: null,
-    args: [
-        { name: "command", type: ArgTypes.CONTENT, required: true }
-    ],
-    // ? Removes comments from a command.
+    type: Types.METHOD,
+    args: [{ name: "command", type: ArgTypes.CONTENT, required: true }],
+    //? Removes comments from a command.
     run: (command: string): string => {
-        const commentIndex = command.indexOf("$c ");
-        if (commentIndex !== -1) {
-            return command.slice(0, commentIndex).trim();
-        }
-        return command.trim();
+        command = command.replace(/\$c\[.*?\]/gs, "").trim();
+        return command.includes("$c ") ? command.split("$c ")[0].trim() : command;
     }
-}
+};