+++ src/modules/commands/init.ts
@@ -1,9 +1,9 @@
 import { commands, text, args as args_1 } from "../Commands";
 import { blue, green, red, yellow } from "../Colors";
 import { Command, Arg } from "../../types/interfaces";
-import TEV from "../../Main";
+import TEV from "../../main";
 import loadArguments from "../Arguments";
-import fs from "node:fs"
+import fs from "node:fs";
 
 const init = async (tev: TEV, args: string[]): Promise<void> => {
     const command: Command = commands.find((command: Command) => command.name == "version")!;