+++ src/structures/maths/index.ts
@@ -1,0 +1,4 @@
+// export * from './angles';
+export * from "./basics";
+export * from "./isNaN";
+export * from "./utils";